move all views to /views/ folder
