+++ src/app/components/home/sheet/sheet.component.ts
@@ -6,7 +6,7 @@ import { ManualTagsService } from '../manual-tags/manual-tags.service';
 import { StarRating, ImageElement } from '../../common/final-object.interface';
 
 import { galleryItemAppear, metaAppear, textAppear } from '../../common/animations';
-import { YearEmission } from '../details/details.component';
+import { YearEmission } from '../views/details/details.component';
 
 export interface TagEmission {
   index: number;
