+++ src/app/components/home/views/clip/clip.component.scss
