+++ src/app/components/home/views/details/details.component.html
