+++ src/app/components/home/views/details/details.component.scss
@@ -54,7 +54,7 @@
 
 .star-rating {
   background-color: transparent;
-  background-image: url("./../../../../assets/logo.png");
+  background-image: url("./../../../../../assets/logo.png");
   background-size: 20px 20px;
   border: 0;
   cursor: pointer;
