+++ src/app/components/home/views/details/details.component.ts
@@ -1,11 +1,11 @@
 import { Component, Input, OnInit, ElementRef, ViewChild, Output, EventEmitter } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
-import { ManualTagsService } from '../manual-tags/manual-tags.service';
+import { ManualTagsService } from '../../manual-tags/manual-tags.service';
 
-import { StarRating, ImageElement } from '../../common/final-object.interface';
+import { StarRating, ImageElement } from '../../../common/final-object.interface';
 
-import { galleryItemAppear } from '../../common/animations';
+import { galleryItemAppear } from '../../../common/animations';
 
 export interface TagEmission {
   index: number;
