+++ src/app/components/home/views/file/file.component.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../variables";
 
 .element-container {
   height: 20px;
