+++ src/app/components/home/views/file/file.component.ts
@@ -1,13 +1,13 @@
 import { Component, Input } from '@angular/core';
 
-import { galleryItemAppear, metaAppear } from '../../common/animations';
-import { ImageElement } from '../../common/final-object.interface';
+import { galleryItemAppear, metaAppear } from '../../../common/animations';
+import { ImageElement } from '../../../common/final-object.interface';
 
 @Component({
   selector: 'app-file-item',
   templateUrl: './file.component.html',
   styleUrls: ['./file.component.scss',
-              '../fonts/icons.scss'],
+              '../../fonts/icons.scss'],
   animations: [ galleryItemAppear,
                 metaAppear ]
 })
