+++ src/app/components/home/views/filmstrip/filmstrip.component.scss
