+++ src/app/components/home/views/full/full.component.scss
