+++ src/app/components/home/views/full/full.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
-import { ImageElement } from '../../common/final-object.interface';
+import { ImageElement } from '../../../common/final-object.interface';
 
-import { galleryItemAppear, metaAppear, textAppear } from '../../common/animations';
+import { galleryItemAppear, metaAppear, textAppear } from '../../../common/animations';
 
 @Component({
   selector: 'app-full-item',
