+++ src/app/components/home/views/thumbnail/preview.component.html
