+++ src/app/components/home/views/thumbnail/preview.component.scss
